requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
bcrypt
requests # For ZhiPu API calls